:art: Remove code duplication
